migrate metadata to blog
